requirements.txt
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=0.20.0